pyproject.toml
[tool.poetry]
name = "task"
version = "0.1.0"
description = ""
authors = ["hubert <hubbertt.k@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
gunicorn = "^23.0.0"
structlog = "^24.4.0"
rich = "^13.9.2"
drf-spectacular = "^0.27.2"



[tool.poetry.group.test.dependencies]
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
